Switch to PEP 440-compliant version numbers
Our version.py script used to return version numbers looking like
`0.1.1-6-gbab2` (`latest tag on the branch`-`number of commits since
that tag`-g`4 hexadecimal digits of the latest commit ID`), which is
not compliant with PEP 440, and thus refused by setuptools 66.0.0.
This new version returns instead `0.1.1.6` (`latest tag on the branch`.`number of commits since that tag`) which is a canonical version as
described in PEP 440.